Do not attempt to rerun failed tasks
Previously we checked if end_time was null, which would include tasks which
had started but not ended. If a task failed, this meant we re-ran the task over
and over again, probably failing every time.

Now we just check if the task has not been previously started, which skips over
failed tasks.